+++ http4k-testing/webdriver/src/main/kotlin/org/http4k/webdriver/Http4kWebDriver.kt
@@ -19,6 +19,7 @@ import org.openqa.selenium.WebElement
 import org.openqa.selenium.WindowType
 import java.net.URL
 import java.nio.file.Paths
+import java.time.Clock
 import java.time.LocalDateTime
 import java.time.ZoneId
 import java.time.ZoneOffset.UTC
@@ -33,9 +34,9 @@ interface Http4KNavigation : Navigation {
     fun to(uri: Uri)
 }
 
-class Http4kWebDriver(initialHandler: HttpHandler) : WebDriver {
+class Http4kWebDriver(initialHandler: HttpHandler, clock: Clock = Clock.systemDefaultZone()) : WebDriver {
     val handler = ClientFilters.FollowRedirects()
-        .then(ClientFilters.Cookies(storage = cookieStorage()))
+        .then(ClientFilters.Cookies(clock, cookieStorage()))
         .then(Filter { next -> { request -> latestUri = request.uri.toString(); next(request) } })
         .then(Filter { next -> { request -> next(request.header("host", latestHost)) } })
         .then(initialHandler)
@@ -69,7 +70,7 @@ class Http4kWebDriver(initialHandler: HttpHandler) : WebDriver {
                 path.startsWith("/") -> Paths.get(path)
                 else -> {
                     val currentPath = currentUrl?.let {
-                        Uri.of(it).path.let { if (it.isEmpty()) "/" else it }
+                        Uri.of(it).path.let { it.ifEmpty { "/" } }
                     } ?: "/"
                     Paths.get(currentPath, path)
                 }